requirements.txt
Django==1.9.4
Markdown==2.6.5
argparse==1.2.1
distribute==0.6.24
django-filter==0.13.0
djangorestframework==3.3.3
wsgiref==0.1.2
django-compressor==2.0
